URPC: Support default values for leafs and leaf-lists Added `default_str` property for yangson values and types, which return the default value string as it appears in the YANG.
Normally yangson sets `default` to a value that matches the value type.
For example default for `type bits` is a tuple of all the bits, and
for `type identityref` is the tuple `(<name_of_identity>, <namespace_of_identity>)`.

This raises an ambiguity issue when a union has `type bits` and `type identity`:
```
identity base-iden;

identity derived-iden {
    base base-iden;
}

leaf a-leaf {
    // default "derived-iden";
    // default "b1 b2";
    type union {
        type bits {
            bit b1;
            bit b2;
        }
        type identityref {
            base base-iden;
        }
    }
}
```

* If the default is "derived-iden" the `default` property will be `('derived-iden', '<namespace>')`

* If the default is "b1 b2" the `default` property will be `('b1', 'b2')`

As the type of `default` is the same, in `generate_urpc` we can't differentiate between the two (in order to rebuild the lexical representation of the default value, to  be set in the YANG metadata default field).